--- datasets_under_construction/wikihow/dummy/all/1.2.0/dummy_data-zip-extracted/dummy_data/all_train.txt
@@ -1,1 +1,0 @@
-test1
